attempted bug fixes

+++ user.c
@@ -2,11 +2,11 @@
 
 extern char __stack_top[];
 
-int syscall(int sysno, int arg0, int arg1, int arg2) {
-    register int a0 __asm__("a0") = arg0;
-    register int a1 __asm__("a1") = arg1;
-    register int a2 __asm__("a2") = arg2;
-    register int a3 __asm__("a3") = sysno;
+int syscall(int sysno, uint64_t arg0, uint64_t arg1, uint64_t arg2) {
+    register uint64_t a0 __asm__("a0") = arg0;
+    register uint64_t a1 __asm__("a1") = arg1;
+    register uint64_t a2 __asm__("a2") = arg2;
+    register uint64_t a3 __asm__("a3") = sysno;
 
     __asm__ __volatile__("ecall"
                          : "=r"(a0)
@@ -37,11 +37,22 @@ __attribute__((noreturn)) void exit(void) {
     for (;;);
 }
 
+// __attribute__((section(".text.start")))
+// __attribute__((naked))
+// void start(void) {
+//     __asm__ __volatile__(
+//         "mv sp, %[stack_top]\n"
+//         "call main\n"
+//         "call exit\n" ::[stack_top] "r"(__stack_top));
+// }
+
 __attribute__((section(".text.start")))
 __attribute__((naked))
 void start(void) {
     __asm__ __volatile__(
         "mv sp, %[stack_top]\n"
         "call main\n"
-        "call exit\n" ::[stack_top] "r"(__stack_top));
+        "call exit\n"
+        :: [stack_top] "r"(__stack_top)
+    );
 }
